Паттерн "Состояния"
Паттерн "Состояния" в сервисе оформления заказа
